+++ R/utils-expr.R
@@ -105,7 +105,7 @@ translate_expr <- function(.data, quo, new_vars = NULL, env, caller) {
 
   # split across() call early
   if (length(expr) > 1 && safe_deparse(expr[[1]]) == "across") {
-    expr <- unpack_across(.data, expr, env)
+    expr <- unpack_across(.data, expr, env, new_vars = new_vars)
   }
 
   # happens because across() calls get split earlier